fix(watchnow): use correct types for rank and delta

+++ projects/api/src/contracts/_internal/response/watchNowRankResponseSchema.ts
@@ -1,7 +1,7 @@
 import { z } from '../z.ts';
 
 export const watchNowRankResponseSchema = z.object({
-  rank: z.number().nullish(),
-  delta: z.number().nullish(),
+  rank: z.number().int().nullish(),
+  delta: z.number().int().nullish(),
   link: z.string().nullish(),
 });
